+++ UVtools.WPF/ErrorLog.cs
@@ -9,14 +9,13 @@ public static class ErrorLog
 {
     private const string Filename = "errors.log";
 
-    public static string FullPath = Path.Combine(UserSettings.SettingsFolder, Filename);
+    public static string FullPath => Path.Combine(CoreSettings.DefaultSettingsFolderAndEnsureCreation, Filename);
 
     public static void AppendLine(string errorType, string text)
     {
         try
         {
-            File.AppendAllText(FullPath,
-                $"[v{About.VersionStr}] ({errorType}) @ {DateTime.Now}: {text}{Environment.NewLine}");
+            File.AppendAllText(FullPath, $"[v{About.VersionStr}] ({errorType}) @ {DateTime.Now}: {text}{Environment.NewLine}");
         }
         catch (Exception exception)
         {